backend/requirements.txt
Flask==3.1.1
google-genai==1.24.0
pymongo==4.13.2
python-dotenv==1.1.1